+++ xbmc/cores/AudioEngine/Engines/ActiveAE/ActiveAEResampleFFMPEG.cpp
@@ -259,13 +259,13 @@ int64_t CActiveAEResampleFFMPEG::GetDelay(int64_t base)
 
 int CActiveAEResampleFFMPEG::GetBufferedSamples()
 {
-  return av_rescale_rnd(swr_get_delay(m_pContext, m_src_rate),
+  return (int)av_rescale_rnd(swr_get_delay(m_pContext, m_src_rate),
                                     m_dst_rate, m_src_rate, AV_ROUND_UP);
 }
 
 int CActiveAEResampleFFMPEG::CalcDstSampleCount(int src_samples, int dst_rate, int src_rate)
 {
-  return av_rescale_rnd(src_samples, dst_rate, src_rate, AV_ROUND_UP);
+  return (int)av_rescale_rnd(src_samples, dst_rate, src_rate, AV_ROUND_UP);
 }
 
 int CActiveAEResampleFFMPEG::GetSrcBufferSize(int samples)
